Add simple command line interface
CLI works line by line and prints a small list of best candidates.
Future work would be to make it work character by character, like a
real keyboard. Making the CLI also exposed an edge case of an empty
prefix. Technically, every word begins with an empty prefix, but it
seems more logical for the application of a keyboard to not complete it.